Rename _timeseries.py to _query.py
