+++ tests/main.py
@@ -42,7 +42,7 @@ import kernel.storage_tests
 import kernel.utils_tests
 import stackdriver.monitoring.metric_tests
 import stackdriver.monitoring.resource_tests
-import stackdriver.monitoring.timeseries_tests
+import stackdriver.monitoring.query_tests
 import stackdriver.monitoring.utils_tests
 import storage.api_tests
 import storage.bucket_tests
@@ -77,7 +77,7 @@ _TEST_MODULES = [
     kernel.utils_tests,
     stackdriver.monitoring.metric_tests,
     stackdriver.monitoring.resource_tests,
-    stackdriver.monitoring.timeseries_tests,
+    stackdriver.monitoring.query_tests,
     stackdriver.monitoring.utils_tests,
     storage.api_tests,
     storage.bucket_tests,
