+++ tests/stackdriver/monitoring/query_tests.py
